+++ idea/testData/quickfix/typeMismatch/changeReturnTypeNoFqNameForCompanionObject.kt
@@ -1,0 +1,11 @@
+// "Change 'Companion.foo' function return type to 'Int'" "true"
+package foo.bar
+
+class A {
+    companion object {
+        fun foo(): String {
+            return <caret>1
+        }
+    }
+}
+